requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
backoff>=2.2.0
python-dotenv>=1.0.0